requirements.txt
gitpython==3.1.41
PyGithub==2.1.1
requests==2.31.0
pandas==2.1.4
scikit-learn==1.3.2
jupyter==1.0.0
